[FLINK-7251][tests] mv test cases to flink-tests
