+++ flink-tests/src/test/java/org/apache/flink/test/operators/lambdas/AllGroupReduceITCase.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -16,22 +16,18 @@
  * limitations under the License.
  */
 
-package org.apache.flink.test.api.java.operators.lambdas;
+package org.apache.flink.test.operators.lambdas;
 
 import org.apache.flink.api.java.DataSet;
 import org.apache.flink.api.java.ExecutionEnvironment;
-import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.test.util.JavaProgramTestBase;
 
 /**
- * IT cases for lambda join functions.
+ * IT cases for lambda allreduce functions.
  */
-public class JoinITCase extends JavaProgramTestBase {
+public class AllGroupReduceITCase extends JavaProgramTestBase {
 
-	private static final String EXPECTED_RESULT = "2,what's really\n" +
-			"2,up really\n" +
-			"1,hello not\n" +
-			"1,hello much\n";
+	private static final String EXPECTED_RESULT = "aaabacad\n";
 
 	private String resultPath;
 
@@ -40,24 +36,19 @@ public class JoinITCase extends JavaProgramTestBase {
 		resultPath = getTempDirPath("result");
 	}
 
-	@SuppressWarnings("unchecked")
 	@Override
 	protected void testProgram() throws Exception {
 		final ExecutionEnvironment env = ExecutionEnvironment.getExecutionEnvironment();
 
-		DataSet<Tuple2<Integer, String>> left = env.fromElements(
-				new Tuple2<Integer, String>(1, "hello"),
-				new Tuple2<Integer, String>(2, "what's"),
-				new Tuple2<Integer, String>(2, "up")
-				);
-		DataSet<Tuple2<Integer, String>> right = env.fromElements(
-				new Tuple2<Integer, String>(1, "not"),
-				new Tuple2<Integer, String>(1, "much"),
-				new Tuple2<Integer, String>(2, "really")
-				);
-		DataSet<Tuple2<Integer, String>> joined = left.join(right).where(0).equalTo(0)
-				.with((t, s) -> new Tuple2<>(t.f0, t.f1 + " " + s.f1));
-		joined.writeAsCsv(resultPath);
+		DataSet<String> stringDs = env.fromElements("aa", "ab", "ac", "ad");
+		DataSet<String> concatDs = stringDs.reduceGroup((values, out) -> {
+			String conc = "";
+			for (String s : values) {
+				conc = conc.concat(s);
+			}
+			out.collect(conc);
+		});
+		concatDs.writeAsText(resultPath);
 		env.execute();
 	}
 
@@ -66,4 +57,3 @@ public class JoinITCase extends JavaProgramTestBase {
 		compareResultsByLinesInMemory(EXPECTED_RESULT, resultPath);
 	}
 }
-
